refactor loader to be ran only once
